refactor(docs-site): colocate specs, drop unused conversion tracking
