+++ apps/docs-site/vitest.config.ts
@@ -10,6 +10,6 @@ export default defineConfig({
   },
   test: {
     environment: 'jsdom',
-    include: ['src/**/__tests__/*.spec.ts'],
+    include: ['src/**/*.spec.ts'],
   },
 })
